feat: split up save by id into separate route
