+++ resources/saves.py
@@ -1,28 +1,22 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint
-from auth.decorators import require_auth
-from schemas import SaveRequestSchema, SaveResponseSchema
-from services.saves_service import get_all_saves, create_save, delete_all_saves, get_save_by_id, delete_save_by_id
+from auth.jwt.decorators import require_auth
+from schemas import SaveResponseSchema
+from services.saves_service import get_all_saves, delete_all_saves
 from flask import g
+import logging
+
+logger = logging.getLogger(__name__)
 
 blp = Blueprint(
-    "Save",
+    "Saves",
     __name__,
     url_prefix="/saves",
-    description="CRUD for save files",
+    description="Managing a collection of save files",
 )
 
 @blp.route("")
 class SaveApi(MethodView):
-    @require_auth
-    @blp.arguments(SaveRequestSchema)
-    @blp.response(201, SaveResponseSchema)
-    def post(self, body):
-        """
-        Create a save file
-        """
-        return create_save(body)
-
     @require_auth
     @blp.response(200, SaveResponseSchema(many=True))
     def get(self):
@@ -39,21 +33,3 @@ class SaveApi(MethodView):
         Delete all save files
         """
         return delete_all_saves()
-
-@blp.route("<string:id>")
-class SaveApi(MethodView):
-    @require_auth
-    @blp.response(200, SaveResponseSchema)
-    def get(self, id):
-        """
-        Get a save file by id
-        """
-        return get_save_by_id(id)
-    
-    @require_auth
-    @blp.response(200)
-    def delete(self, id):
-        """
-        Delete a save file by id
-        """
-        return delete_save_by_id(id)
