+++ schemas/saves.py
@@ -7,4 +7,4 @@ class SaveResponseSchema(Schema):
     id = fields.Str(dump_only=True)
     location = fields.Str(required=True)
     party = fields.List(fields.Str(dump_only=True))
-    party_lead = fields.Dict()
+    party_lead = fields.Dict(dump_only=True)
